Platformer Minigame Prototype
This uses very basic shapes for the prototype of the mini game. All that needs to be added is the actual assets, the goal object in order to go to next mini game, and the alteration. The alteration could be the entire screen spinning. Also needs to add jump limit